requirements.txt
flask
gunicorn
openai
faiss-cpu
numpy
python-dotenv
pdfplumber
python-docx
